django-mysql-docker-app/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
mysqlclient==2.1.1
python-dotenv==1.0.0